feat(auth): taken accountType in account to define users granted access

+++ src/Controller/LoginAttemptController.php
@@ -27,12 +27,25 @@ class LoginAttemptController extends AbstractController
     {
         $params = $request->getParsedBody();
         $user = $this->auth->login($params['email'], $params['password']);
-        if ($user) {
+        if ($user && true === $user->getAccountStatus()) {
+            if ('user' === $user->getAccountType()) {
+                $path = $this->router->generateUri('site.home', []);
+
+                return new RedirectResponse($path);
+            }
             $path = $this->session->get('auth.redirect')
                 ?: $this->router->generateUri('admin.dashboard', []);
             $this->session->delete('auth.redirect');
 
             return new RedirectResponse($path);
+        }
+        if ($user && false === $user->getAccountStatus()) {
+            (new FlashService($this->session))
+                ->error(
+                    'Votre compte n\'est pas encore activé ! Patientez SVP'
+                );
+
+            return $this->router->redirect('auth.login');
         }
         (new FlashService($this->session))
             ->error('Identifiant ou mot de passe incorrect');
